Refactor index.js to correct PORT environment variable casing, add FETCH_TIMEOUT_MS constant in config.js, and improve error handling in newbalanceController and scraperService
